chore(EMS-1966): split up UI API calls
